requirements.txt
aiolimiter==1.1.0
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
filelock==3.13.4
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.26.0
httpx-sse==0.3.1
huggingface-hub==0.22.2
ibm-generative-ai==2.3.0
idna==3.7
multiprocess==0.70.16
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.29
tokenizers==0.19.1
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0
